# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TutorTopia API",
    "description": "This is a simple API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://tutortopia-7d536-default-rtdb.firebaseio.com",
      "description": "TutorTopia API Auto Mocking"
    },
    {
      "url": "https://randomurl.com",
      "description": "Stand in URL until host is created"
    }
  ],
...
  "paths": {
...
    "/user/availability/{userId}": {
      "get": {
        "tags": [
          "Availability"
        ],
        "summary": "Gets availability for the user",
        "description": "Finds user profile by id",
        "operationId": "searchUserAvailability",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Availability"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          }
        }
      },
      "put": {
        "tags": [
          "Availability"
        ],
        "summary": "Update availability by ID",
        "description": "Update the availability related to user ID\n",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Availability"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          }
        }
      },
      "post": {
        "tags": [
          "Availability"
        ],
        "summary": "adds a new availability",
        "description": "Adds an availability for user to the system",
        "operationId": "addUserAvailability",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "availability created"
          },
          "400": {
            "description": "invalid input, object invalid"
          }
        }
      },
      "delete": {
        "tags": [
          "Availability"
        ],
        "summary": "Delete availability by user ID",
        "description": "Delete the availability linked to user ID\n",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Availability"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Availability": {
        "type": "object",
        "properties": {
          "tutorId": {
            "type": "string"
          },
          "weekly": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "exceptions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}